fixed styling for hover and added linking to segment text
