+++ src/core/svg/SVGStyles.ts
@@ -8,5 +8,6 @@ export const ModeCSS = {
   BLUE: "stroke-blue-500 stroke-[3px] fill-none opacity-1",
   ACTIVETEXT: "fill-violet-500 stroke-violet-500 stroke-[1px]",
   PINNED: "stroke-violet-500 stroke-[4px] pinnedSVG",
-  DIAGRAMHOVER: "stroke-[#8b5cf6] stroke-[3px] drop-shadow-[0_0_5px_#8b5cf6]",
+  DIAGRAMHOVER: "stroke-violet-600 stroke-[3px] drop-shadow-[0_0_5px_#8b5cf6]",
+  DIAGRAMCLICKTEXT: "text-violet-500 font-black",
 };
